adatbazisban el volt irva Available Avaiblere es valamiert meg az nem jo majd kesobb meg ra kell nezni

+++ HotelGuru/HotelGuru.DataContext/Entities/Invoice.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Linq;
-using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -11,12 +9,11 @@ namespace HotelGuru.DataContext.Entities
     public class Invoice
     {
         public int Id { get; set; }
-        public int Price { get; set; }
-        public int ExtraPrice { get; set; }
-        public int TotalPrice => Price + ExtraPrice;
+        public decimal Price { get; set; }
+        public decimal ExtraPrice { get; set; }
+        public decimal TotalPrice => Price + ExtraPrice;
 
         public int ReservationId { get; set; }
         public Reservation Reservation { get; set; } = null!;
     }
-
 }
